fix return value type

+++ ttsd-aa-repository/src/main/java/com/tuotiansudai/repository/mapper/InvestMapper.java
@@ -295,7 +295,7 @@ public interface InvestMapper {
                                                   @Param(value = "loanStatus") LoanStatus loanStatus);
 
 
-    int findFirstInvestAmountByLoginName(@Param(value = "loginName") String loginName,
+    long findFirstInvestAmountByLoginName(@Param(value = "loginName") String loginName,
                                 @Param(value = "startTime") Date startTime,
                                 @Param(value = "endTime") Date endTime);
 }
